feat(Clinvar Submissions): more dry Separates SubmissionsList component
